move adapter into a sub package
